Fixed our bug
We added an optional heightConstraint property that can be updated with
the height of the UITextView. This height gets calculated with
UITextView’s sizeThatFits method in its delegate textViewDidChange
method.

We also set a default height that we can initialize the text view with,
and fall back to if the calculated size is too small.